% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_presences.R
\name{clean_presences}
\alias{clean_presences}
\title{Clean presences}
\usage{
clean_presences(DF)
}
\arguments{
\item{DF}{A data frame of presence records}
}
\value{
A data frame of cleaned presence records
}
\description{
Cleans presence data
}
\examples{
Presences <- SDMWorkflows::GetOccs(Species = c("Abies concolor"), WriteFile = FALSE, limit = 500)
Cleaned <- clean_presences(Presences[[1]])
}
