% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean_Taxa.R
\name{Clean_Taxa}
\alias{Clean_Taxa}
\title{Clean taxa using Taxize and rgbif}
\usage{
Clean_Taxa(Taxons, WriteFile = F, Species_Only = T)
}
\arguments{
\item{Taxons}{Vector of taxa to be cleaned.}

\item{WriteFile}{logical if False (default) only returns a
data frame, if TRUE will generate a folder (Results in the
Working directory with a csv of the results)}

\item{Species_Only}{logical, if TRUE (default) only species will be
returned, if FALSE, it will return the highest possible taxonomic
resolution}
}
\value{
A data frame with the cleaned taxa and their scores.
}
\description{
This function cleans a vector of taxa using Taxize and rgbif
}
\examples{

Cleaned <- Clean_Taxa(Taxons = c("Canis lupus", "C. lupus"))

}
\references{
Chamberlain, Scott T., et al. "Taxize: an R package for taxonomic
lookup and manipulation." Methods in Ecology and Evolution 6.12
(2015): 1451-1456.
Chamberlain S, Barve V, Mcglinn D, Oldoni D, Desmet P, Geffert L, Ram K
(2023). rgbif: Interface to the Global Biodiversity Information
Facility API_ R package version 3.7.4
}
