% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean_Taxa.R
\name{Clean_Taxa_Taxize}
\alias{Clean_Taxa_Taxize}
\title{Clean taxa using Taxize}
\usage{
Clean_Taxa_Taxize(Taxons)
}
\arguments{
\item{Taxons}{Vector of taxa to be cleaned.}
}
\value{
A data frame with the cleaned taxa and their scores.
}
\description{
This function cleans a vector of taxa using Taxize.
}
\examples{
Clean_Taxa_Taxize(Taxons = c("Canis lupus", "C. lupus"))

}
\references{
Chamberlain, Scott T., et al. "Taxize: an R package for taxonomic
lookup and manipulation." Methods in Ecology and Evolution 6.12
(2015): 1451-1456.
}
