% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convex_20.R
\name{Convex_20}
\alias{Convex_20}
\title{Convex_20}
\usage{
Convex_20(
  DF,
  lon = "decimalLongitude",
  lat = "decimalLatitude",
  proj = "+proj=longlat +datum=WGS84 +no_defs"
)
}
\arguments{
\item{DF}{The dataframe containing the coordinates.}

\item{lon}{The name of the longitude column in the dataframe.}

\item{lat}{The name of the latitude column in the dataframe.}

\item{proj}{The projection of the coordinates.}
}
\value{
A polygon representing the expanded convex hull by 20\%.
}
\description{
Creates a 20\% expanded convex hull from a set of coordinates.
}
\examples{

DF <- data.frame(decimalLongitude = c(23.978543, 23.785003, 11.485,  -2.054027, 12.9069), decimalLatitude = c(38.088876, 60.238213, 48.165, 53.33939, 56.80782))

Convex_20(DF, lon = "decimalLongitude", lat = "decimalLatitude", proj = "+proj=longlat +datum=WGS84 +no_defs")

}
