% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean_Taxa.R
\name{Clean_Taxa_rgbif}
\alias{Clean_Taxa_rgbif}
\title{Clean Taxa from GBIF}
\usage{
Clean_Taxa_rgbif(Cleaned_Taxize)
}
\arguments{
\item{Cleaned_Taxize}{a data frame containing the cleaned taxonomic list from function Clean_Taxa_Taxize}
}
\value{
A csv file containing the cleaned taxonomic list
}
\description{
Clean the taxonomic list using GBIF
}
\examples{
Cleaned_Taxize <- Clean_Taxa_Taxize(Taxons = c("Canis lupus", "C. lupus"))
Clean_Taxa_rgbif(Cleaned_Taxize)
}
\references{
Chamberlain S, Barve V, Mcglinn D, Oldoni D, Desmet P, Geffert L, Ram K
(2023). rgbif: Interface to the Global Biodiversity Information
Facility API_ R package version 3.7.4,
}
