% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_presence.R
\name{GetOccs}
\alias{GetOccs}
\title{Get occurrence data}
\usage{
GetOccs(Species, WriteFile = T, continent = NULL, limit = 10000)
}
\arguments{
\item{Species}{A data frame containing the species to query.}

\item{WriteFile}{Logical. If \code{TRUE}, the occurrence data will be written to the \code{Occs} folder. If \code{FALSE}, the occurrence data will be returned in a list.}
}
\value{
If \code{WriteFile = TRUE}, this function does not return anything. If \code{WriteFile = FALSE}, a list containing the occurrence data for each species is returned.
}
\description{
This function uses the \code{rgbif} package to get occurrence data from the Global Biodiversity Information Facility (GBIF) API.
}
\examples{
# Get occurrence data for species in FinalSpeciesList
Presences <- SDMWorkflows::GetOccs(Species = c("Abies concolor", "Canis lupus"), WriteFile = FALSE)
}
